Push CanBeRepaired check into base job
If HEAD is currently pointing to a ref that we're about to fix, we are
in a sense updating where HEAD is. We should prevent operations that
change (or might change) the position of HEAD if we're in the middle of
a repository operation such as rebase, cherry-pick etc.
